requirements.txt
backcall==0.2.0
click==7.1.2
cmake==3.18.4.post1
decorator==4.4.2
Django==2.2.17
djangorestframework==3.12.2
dlib==19.21.1
face-recognition==1.3.0
face-recognition-models==0.3.0
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
numpy==1.19.4
opencv-python==4.4.0.46
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
ptyprocess==0.6.0
Pygments==2.7.3
pytz==2020.4
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
